+++ obsidianlink/env/__init__.py
@@ -2,6 +2,7 @@
 
 from obsidianlink.env.actions import Action, ActionType
 from obsidianlink.env.environment import Environment, Observation
+from obsidianlink.env.l1_scene import L1ControlledEnv
 from obsidianlink.env.minerl import MineRLEnvironment
 from obsidianlink.env.scene import ControlledSceneEnv
 
@@ -10,6 +11,7 @@ __all__ = [
     "ActionType",
     "ControlledSceneEnv",
     "Environment",
+    "L1ControlledEnv",
     "MineRLEnvironment",
     "Observation",
 ]